# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: |
    SubNation Digital Subscriptions Platform API.

    ## Authentication (round-93 / 93-A8 F-1 — contract truth)

    Two session mechanisms, both documented as security schemes and
    attached per operation family:

    - **User routes** (auth, orders, wallet, cart, coupons/validate,
      loyalty, referrals, support, notifications): an `auth_token`
      httpOnly cookie (primary) OR an `Authorization: Bearer <token>`
      header (fallback).
    - **Admin routes** (/admin/*, /coupons/admin*, the copilot family):
      an `admin_token` httpOnly cookie (8h, issued by POST
      /admin/login) OR its Bearer equivalent. POST /admin/login itself
      is the credential endpoint and needs no prior session.

    ## CSRF gate (applies to every state-changing request)

    Every POST/PUT/PATCH/DELETE under /api — regardless of cookie or
    Bearer authentication — must carry an Origin (or Referer) header
    matching the deployment's allow-list (CSRF_ALLOWED_ORIGINS /
    APP_ORIGINS / APP_URL); otherwise the request is rejected with
    403 FORBIDDEN before the route runs. Requests with neither Origin
    nor Referer are also rejected. Exemptions:
    /api/auth/firebase/session, /api/auth/firebase/refresh, /api/cwv,
    /api/webhook/*, and /health. Non-browser clients must therefore
    send a browser-like Origin header on mutations.

    ## Deployment-conditional behavior

...
servers:
  - url: /api
...
security:
  - cookieAuth: []
    bearerAuth: []
...
paths:
...
  /coupons/admin:
    get:
      operationId: listAdminCoupons
      tags: [coupons]
      security:
        - adminCookieAuth: []
          adminBearerAuth: []
      summary: List all coupons (requireAdmin + finance scope)
      description: Newest first; includes inactive/soft-deactivated rows.
      responses:
        "200":
          description: All coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing finance permission scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCoupon
      tags: [coupons]
      security:
        - adminCookieAuth: []
          adminBearerAuth: []
      summary: Create a coupon (requireAdmin + finance scope)
      description: |
        zod-validated (CreateCouponBody in routes/coupons.ts). `code` is
        uppercased before insert. Percentage coupons must have value ≤ 100
        (checked after parse); fixed coupons are bounded at 10,000 LYD.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "400":
          description: Validation error, or percentage value above 100 (INVALID_DATA)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing finance permission scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Coupon code already exists (ALREADY_EXISTS)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /coupons/admin/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      security:
        - adminCookieAuth: []
          adminBearerAuth: []
      summary: Update coupon mutable fields (requireAdmin + finance scope)
      description: Strict zod body — unknown keys are a 400.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchCouponBody"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "400":
          description: Invalid id or body (INVALID_DATA)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing finance permission scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Coupon not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      security:
        - adminCookieAuth: []
          adminBearerAuth: []
      summary: Soft-deactivate a coupon (requireAdmin + finance scope)
      description: Sets is_active=false — the row is kept for audit history.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid id (INVALID_DATA)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing finance permission scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Coupon not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Loyalty ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    # Round-93 (93-A8 F-1): the spec previously documented zero security
    # definitions while the middleware reality is cookie-first with a
    # Bearer fallback on user routes, and a separate admin cookie (8h)
    # plus Bearer on admin routes. Both alternatives are expressed per
    # operation via `security: [{ schemeA: [], schemeB: [] }]` (same
    # object = OR).
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: |
        User session — httpOnly cookie set by the auth routes (Google /
        Telegram / WhatsApp-OTP sign-in). SameSite=None in production so
        the Vercel-hosted SPA can call the API cross-site. Primary
        mechanism for browser clients.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        User-session Bearer fallback — `Authorization: Bearer <token>`
        accepted wherever cookieAuth is (requireUser checks the cookie
        first, then the header).
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: |
        Admin session — separate httpOnly cookie (8h) issued by POST
        /admin/login (and verify-2fa when TOTP is enabled).
    adminBearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin-session Bearer alternative — the token returned in the
        POST /admin/login JSON body, accepted wherever adminCookieAuth
        is (requireAdmin accepts either).
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # Standard error envelope. Every non-2xx /api response carries
    # {error, code} — createErrorResponse (backend/src/lib/errors.ts), the
    # rate limiters, mapServiceErrorToCode and the global handler all
    # produce it. Zod-backed ValidationError bodies may additionally carry
    # `details` (zod issue arrays) or other extra keys, but the guaranteed
    # contract is {error, code}.
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          $ref: "#/components/schemas/ErrorCode"
        details:
          type: object
          additionalProperties: true
          description: |
            Optional structured context — zod `issues` on validation
            failures, `{field}` hints on flash-sale validation errors,
            `{skipped_duplicates}` on inventory uploads.
      required: [error, code]

    # Machine-readable error codes — mirrors the shared ErrorCode enum
    # in @workspace/error-codes (shared/error-codes/src/index.ts), the
    # single source both backend and frontend re-export. The copilot
    # preview/execution family (round-92 B3/F-05) closes the gap where
    # backend/src/routes/admin/copilot/* returned codes outside the enum.
    ErrorCode:
      type: string
      enum:
        - INVALID_DATA
        - INVALID_PHONE
        - INVALID_PASSWORD_LENGTH
        - INVALID_PASSWORD_WEAK
        - INVALID_OTP
        - INVALID_CREDENTIAL
        - UNAUTHORIZED
        - INVALID_TOKEN
        - SESSION_EXPIRED
        - ACCOUNT_LOCKED
        - ACCOUNT_NOT_FOUND
        - PHONE_ALREADY_REGISTERED
        - FEATURE_DISABLED
        - FORBIDDEN
        - INSUFFICIENT_PERMISSIONS
        - NOT_FOUND
        - ALREADY_EXISTS
        - CONFLICT
        - OUT_OF_STOCK
        - PRODUCT_UNAVAILABLE
        - INSUFFICIENT_BALANCE
        - INVALID_AMOUNT
        - TOPUP_LIMIT_EXCEEDED
        - ORDER_NOT_FOUND
        - ORDER_ALREADY_COMPLETED
        - ORDER_CANNOT_CANCEL
        - GOOGLE_TOKEN_INVALID
        - GOOGLE_VERIFICATION_FAILED
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
        - RATE_LIMITED
        - IDEMPOTENCY_IN_FLIGHT
        - IDEMPOTENCY_KEY_REUSE
        - COPILOT_INVALID_INPUT
        - COPILOT_LLM_ERROR
        - COPILOT_NO_ADMIN_SESSION
        - COPILOT_BAD_METHOD
        - COPILOT_PREVIEW_NOT_FOUND
        - COPILOT_PREVIEW_CONSUMED
        - COPILOT_PREVIEW_EXPIRED
        - COPILOT_HANDOFF_REQUIRED
        - COPILOT_HIGH_RISK_DISABLED
        - COPILOT_STALE_RECORD
        - COPILOT_UNEXPECTED_STATE
        - COPILOT_EXECUTE_FAILED
        - COPILOT_NOT_HIGH_RISK
        - COPILOT_FIRST_CONFIRM_MISSING
        - COPILOT_COOLDOWN_NOT_ELAPSED
        - COPILOT_LLM_UNAVAILABLE
        - COPILOT_OUT_OF_SCOPE
        - COPILOT_INVALID_FLAGS
        # Round-93 (93-A8 F-8/F-15): two emitted codes that previously
        # lived outside the whole code contract — the copilot rate
        # limiter's 429 body and the CSRF gate's fail-closed
        # misconfiguration branch. The enum stays exhaustive over what
        # the backend actually emits.
        - COPILOT_RATE_LIMITED
        - CSRF_CONFIG
        # Round-94 (A5-05): the last two emitted-but-unenumerated codes —
        # the copilot phase-gate 503 (requireCopilotPhase) and the
        # outbound secret-scan 502 abort (ask/draft routes).
        - COPILOT_PHASE_DISABLED
        - COPILOT_SECRET_LEAK

    # 409 body returned by the idempotency middleware
    # (backend/src/middlewares/idempotency.ts): the same Idempotency-Key
    # re-sent with a different body (IDEMPOTENCY_KEY_REUSE) or the same
    # key while the original request is still executing
    # (IDEMPOTENCY_IN_FLIGHT). `message` is a legacy alias kept for older
    # clients — `error` is the canonical field.
...
    Coupon:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
        min_order_amount:
          type: number
        max_uses:
          type: ["integer", "null"]
        used_count:
          type: integer
        expires_at:
          type: ["string", "null"]
        is_active:
          type: boolean
        description:
          type: ["string", "null"]
        created_at:
          type: string
      required: [id, code, type, value, used_count, is_active, created_at]
    CreateCouponBody:
      type: object
      description: |
        `code` is uppercased before insert. Percentage coupons must have
        value ≤ 100 (checked after parse with a specific message); fixed
        values are bounded at 10,000 LYD — a direct wallet-debit magnitude
        at checkout.
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 40
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
          exclusiveMinimum: 0
          maximum: 10000
        min_order_amount:
          type: number
          minimum: 0
          maximum: 1000000
          description: Defaults to 0.
        max_uses:
          type: ["integer", "null"]
          minimum: 1
          maximum: 1000000
        expires_at:
          type: ["string", "null"]
          description: ISO date or date-time string; null = never expires.
        description:
          type: ["string", "null"]
          maxLength: 200
      required: [code, type, value]
    PatchCouponBody:
      type: object
      description: Strict body — unknown keys are a 400 (zod .strict()).
      properties:
        is_active:
          type: boolean
        max_uses:
          type: ["integer", "null"]
          minimum: 1
          maximum: 1000000
        expires_at:
          type: ["string", "null"]
          description: ISO date or date-time string; null = never expires.
        description:
          type: ["string", "null"]
          maxLength: 200

    # ── Loyalty ──────────────────────────────────────────────────────────────
...